refactor(mainmenu): userinterface 패키지 하위에 mainmenu 패지키로 분리
