+++ src/main/java/subway/userinterface/mainmenu/MainViewInputController.java
@@ -1,4 +1,4 @@
-package subway.userinterface;
+package subway.userinterface.mainmenu;
 
 import subway.util.InputValidator;
