+++ src/main/java/subway/userinterface/mainmenu/MainViewOutputController.java
@@ -1,4 +1,6 @@
-package subway.userinterface;
+package subway.userinterface.mainmenu;
+
+import subway.userinterface.Menu;
 
 import java.util.Map;
